Version 0.1.0
This is the first of the Version 0.1.x releases. It includes XMPP. Rather than removing the SIP SIMPLE messaging system, this version adds XMPP on as an option, allowing for slightly easier transition from the old version. With XMPP the buddies are now stored in the cloud, but all the information around call details, recordings etc are still saved locally. The XMPP is very basic and over covers a small feature set of what XMPP can do. The idea is to iron out the bugs for the core XMPP features and then get to the new XMPP features, like file and image sharing etc.

Other than the addition of XMPP, this version includes:
-	Allows multiple calls to be ringing inbound, at the same time.
-	Allows for multiple outbound call to the same buddy to be made.
-	A very basic Early Media implementation (without Rel100).
-	An option to limit outbound video bandwidth.
And quite a few bug fixes.

As always please feel free to test, and share your results.