Version 0.2.2
Added ConfbridgeMute and ConfbridgeUnmute events, although these are not very accurate, as Asterisk is confused about this, and doesn't send the correct state.

Added extraHeaders parameter to the DialByLine function

Added runtime configurations. These can be parsed from the html page and are not written to the database. Allows for better provisioning, and opening the phone with different configurations within the same domain.